Align info icon to title better
