+++ frontend/src/lib/components/Cards/CardMeta.tsx
@@ -1,17 +1,15 @@
 import './CardMeta.scss'
 
 import clsx from 'clsx'
-import React, { useCallback, useEffect, useRef, useState } from 'react'
+import React from 'react'
 import { Transition } from 'react-transition-group'
 
-import { IconInfo, IconPieChart } from '@posthog/icons'
+import { IconPieChart } from '@posthog/icons'
 
-import { EditableField } from 'lib/components/EditableField/EditableField'
 import { useResizeObserver } from 'lib/hooks/useResizeObserver'
 import { IconSubtitles, IconSubtitlesOff } from 'lib/lemon-ui/icons'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { More } from 'lib/lemon-ui/LemonButton/More'
-import { Popover } from 'lib/lemon-ui/Popover'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { inStorybook, inStorybookTestRunner } from 'lib/utils'
 
@@ -71,11 +69,6 @@ export function CardMeta({
     onMouseDown,
     samplingFactor,
     extraControls,
-    metaDescription,
-    metaTitle,
-    popoverTopHeading,
-    metaDescriptionText,
-    onMetaSave,
 }: CardMetaProps): JSX.Element {
     const { ref: primaryRef, width: primaryWidth } = useResizeObserver()
     const { ref: detailsRef, height: detailsHeight } = useResizeObserver()
@@ -96,41 +89,6 @@ export function CardMeta({
         inStorybook() ||
         inStorybookTestRunner() ||
         (!!primaryWidth && primaryWidth > 480 && controlsAvailableSpace >= neededWidth)
-
-    const [popoverVisible, setPopoverVisible] = useState(false)
-    const [pinned, setPinned] = useState(false)
-    const hoverTimerRef = useRef<ReturnType<typeof setTimeout>>()
-
-    useEffect(() => () => clearTimeout(hoverTimerRef.current), [])
-
-    const clearHoverTimer = (): void => clearTimeout(hoverTimerRef.current)
-
-    const showDetails = useCallback(() => {
-        clearHoverTimer()
-        hoverTimerRef.current = setTimeout(() => setPopoverVisible(true), 300)
-    }, [])
-
-    const hideDetails = useCallback(() => {
-        clearHoverTimer()
-        hoverTimerRef.current = setTimeout(() => setPopoverVisible(false), 800)
-    }, [])
-
-    const handleClickInfo = useCallback(
-        (e: React.MouseEvent) => {
-            e.stopPropagation()
-            clearHoverTimer()
-            const newPinned = !pinned
-            setPinned(newPinned)
-            setPopoverVisible(newPinned)
-        },
-        [pinned]
-    )
-
-    const handleClickOutside = useCallback(() => {
-        clearHoverTimer()
-        setPinned(false)
-        setPopoverVisible(false)
-    }, [])
 
     return (
         <div
@@ -165,79 +123,7 @@ export function CardMeta({
                                         ))}
                                 </div>
                             </div>
-                            <div className="flex items-start gap-1">
-                                <div className="overflow-hidden min-w-0">{meta}</div>
-                                <Popover
-                                    visible={popoverVisible}
-                                    placement="bottom"
-                                    showArrow
-                                    onClickOutside={handleClickOutside}
-                                    onMouseEnterInside={showDetails}
-                                    onMouseLeaveInside={pinned ? undefined : hideDetails}
-                                    overlay={
-                                        <div
-                                            className="p-4 max-w-md space-y-2"
-                                            onMouseDown={(e) => e.stopPropagation()}
-                                        >
-                                            {(popoverTopHeading ?? topHeading) && (
-                                                <h5 className="uppercase text-xs font-bold text-muted tracking-wide m-0">
-                                                    {popoverTopHeading ?? topHeading}
-                                                </h5>
-                                            )}
-                                            {onMetaSave ? (
-                                                <>
-                                                    <EditableField
-                                                        name="title"
-                                                        value={metaTitle || ''}
-                                                        onSave={(value) => onMetaSave({ name: value })}
-                                                        placeholder="Untitled"
-                                                        saveOnBlur
-                                                        clickToEdit
-                                                        compactButtons
-                                                        compactIcon
-                                                        className="font-semibold text-sm"
-                                                        data-attr="insight-card-title"
-                                                    />
-                                                    <EditableField
-                                                        name="description"
-                                                        value={metaDescriptionText || ''}
-                                                        onSave={(value) => onMetaSave({ description: value })}
-                                                        placeholder="Enter description (optional)"
-                                                        saveOnBlur
-                                                        clickToEdit
-                                                        multiline
-                                                        markdown
-                                                        compactButtons
-                                                        compactIcon
-                                                        className="text-xs w-full"
-                                                        data-attr="insight-card-description"
-                                                    />
-                                                </>
-                                            ) : (
-                                                <>
-                                                    {metaTitle && <p className="font-semibold m-0">{metaTitle}</p>}
-                                                    {metaDescription}
-                                                </>
-                                            )}
-                                            {metaDetails}
-                                        </div>
-                                    }
-                                >
-                                    <div
-                                        className="pt-0.5"
-                                        onMouseEnter={showDetails}
-                                        onMouseLeave={pinned ? undefined : hideDetails}
-                                    >
-                                        <LemonButton
-                                            icon={<IconInfo />}
-                                            size="small"
-                                            noPadding
-                                            data-attr="card-meta-info"
-                                            onClick={handleClickInfo}
-                                        />
-                                    </div>
-                                </Popover>
-                            </div>
+                            {meta}
                         </>
                     ) : (
                         <>
